Testing-Projects-main/requirements.txt
python-telegram-bot
requests
pillow
aiohttp